requirements.txt
###############################################################
# Main App
gradio>=5.2     # main UI
colorlog        # Colored logging output
Pillow          # PIL images
numpy           # Image Manipulation (SKIP_AI)
dotenv          # Load environment variables for config from file
apscheduler     # Background scheduler for Gradio Server to count active sessions or free memory by unloading model

###############################################################
# Analytics
user-agents     # to parse user agents and split into 
prometheus-client    # for metrics collection and exposition

###############################################################
# AI
diffusers>=0.32         # for stable diffusion general
transformers>=4.49      # for image generation
accelerate              #for low cpu
torch>=2.6              # for gpu support and ai operations
xformers                # memory optimization
protobuf                # required by pipeline
sentencepiece           # required by tokenizer

###############################################################
# Image Analyzer
insightface             # to detect faces in uploaded images
onnxruntime             # required by insightface but not automatically installed          
opencv-python           # to prepare PIL images for face detection
nudenet                 # nsfw detector
###############################################################
# Prompt Analyzer
langchain-ollama        # to query the system
ollama                  # to pull the correct models